+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/HttpRequestMethodNotSupportedProblemResolver.java
@@ -19,6 +19,7 @@ package io.github.problem4j.spring.web.resolver;
 import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -46,7 +47,7 @@ public class HttpRequestMethodNotSupportedProblemResolver extends AbstractProble
    * @since 1.2.0
    */
   public HttpRequestMethodNotSupportedProblemResolver() {
-    this(ProblemFormat.identity());
+    super(HttpRequestMethodNotSupportedException.class);
   }
 
   /**
@@ -55,7 +56,12 @@ public class HttpRequestMethodNotSupportedProblemResolver extends AbstractProble
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link
+   *     #HttpRequestMethodNotSupportedProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public HttpRequestMethodNotSupportedProblemResolver(ProblemFormat problemFormat) {
     super(HttpRequestMethodNotSupportedException.class, problemFormat);
   }