+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/MethodValidationProblemResolver.java
@@ -22,8 +22,10 @@ import static io.github.problem4j.spring.web.parameter.ViolationSupport.VALIDATI
 import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import io.github.problem4j.spring.web.parameter.DefaultMethodValidationResultSupport;
 import io.github.problem4j.spring.web.parameter.MethodValidationResultSupport;
+import io.github.problem4j.spring.web.parameter.MethodValidationResultSupportAware;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -47,12 +49,18 @@ import org.springframework.validation.method.MethodValidationException;
  * extension populated via {@link MethodValidationResultSupport} (one entry per violated parameter /
  * return value).
  *
+ * <p>When used as a Spring bean, in addition to the {@link ProblemFormat} injected via {@link
+ * AbstractProblemResolver}, the {@link MethodValidationResultSupport} is assigned after
+ * construction by {@link ProblemBeanPostProcessor} through {@link
+ * #setMethodValidationResultSupport(MethodValidationResultSupport)}.
+ *
  * @see jakarta.validation.ConstraintViolationException
  * @since 1.2.0
  */
-public class MethodValidationProblemResolver extends AbstractProblemResolver {
+public class MethodValidationProblemResolver extends AbstractProblemResolver
+    implements MethodValidationResultSupportAware {
 
-  private final MethodValidationResultSupport methodValidationResultSupport;
+  private MethodValidationResultSupport methodValidationResultSupport;
 
   /**
    * Creates a new {@link MethodValidationProblemResolver} with default problem format.
@@ -60,7 +68,8 @@ public class MethodValidationProblemResolver extends AbstractProblemResolver {
    * @since 1.2.0
    */
   public MethodValidationProblemResolver() {
-    this(ProblemFormat.identity());
+    super(MethodValidationException.class);
+    this.methodValidationResultSupport = new DefaultMethodValidationResultSupport();
   }
 
   /**
@@ -68,7 +77,10 @@ public class MethodValidationProblemResolver extends AbstractProblemResolver {
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns collaborators after
+   *     construction; use {@link #MethodValidationProblemResolver()}
    */
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public MethodValidationProblemResolver(ProblemFormat problemFormat) {
     this(problemFormat, new DefaultMethodValidationResultSupport());
   }
@@ -80,10 +92,26 @@ public class MethodValidationProblemResolver extends AbstractProblemResolver {
    * @param problemFormat the problem format to use
    * @param methodValidationResultSupport the support for extracting validation results
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns collaborators after
+   *     construction; use {@link #MethodValidationProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public MethodValidationProblemResolver(
       ProblemFormat problemFormat, MethodValidationResultSupport methodValidationResultSupport) {
     super(MethodValidationException.class, problemFormat);
+    this.methodValidationResultSupport = methodValidationResultSupport;
+  }
+
+  /**
+   * Replaces the {@link MethodValidationResultSupport} used by this resolver.
+   *
+   * @param methodValidationResultSupport the method validation result support to use
+   * @since 3.1.0
+   */
+  @Override
+  public void setMethodValidationResultSupport(
+      MethodValidationResultSupport methodValidationResultSupport) {
     this.methodValidationResultSupport = methodValidationResultSupport;
   }