+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/MissingRequestValueProblemResolver.java
@@ -41,6 +41,7 @@ import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemBuilder;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import java.util.Locale;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -67,7 +68,7 @@ public class MissingRequestValueProblemResolver extends AbstractProblemResolver 
    * @since 1.2.0
    */
   public MissingRequestValueProblemResolver() {
-    this(ProblemFormat.identity());
+    super(MissingRequestValueException.class);
   }
 
   /**
@@ -75,7 +76,11 @@ public class MissingRequestValueProblemResolver extends AbstractProblemResolver 
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link #MissingRequestValueProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public MissingRequestValueProblemResolver(ProblemFormat problemFormat) {
     super(MissingRequestValueException.class, problemFormat);
   }