+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/MultipartProblemResolver.java
@@ -19,6 +19,7 @@ package io.github.problem4j.spring.web.resolver;
 import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -41,7 +42,7 @@ public class MultipartProblemResolver extends AbstractProblemResolver {
    * @since 1.2.0
    */
   public MultipartProblemResolver() {
-    this(ProblemFormat.identity());
+    super(MultipartException.class);
   }
 
   /**
@@ -49,7 +50,11 @@ public class MultipartProblemResolver extends AbstractProblemResolver {
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link #MultipartProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public MultipartProblemResolver(ProblemFormat problemFormat) {
     super(MultipartException.class, problemFormat);
   }